move text blueprint from yafowil.common to yafowil.text
